+++ packages/editor-core/src/editor/index.tsx
@@ -100,11 +100,13 @@ export default class Editor extends React.Component<EditorProps, State> {
       collabEditProvider,
       activityProvider,
       presenceProvider,
-      macroProvider
+      macroProvider,
+      legacyImageUploadProvider
     } = props;
     this.providerFactory.setProvider('emojiProvider', emojiProvider);
     this.providerFactory.setProvider('mentionProvider', mentionProvider);
     this.providerFactory.setProvider('mediaProvider', mediaProvider);
+    this.providerFactory.setProvider('imageUploadProvider', legacyImageUploadProvider);
     this.providerFactory.setProvider('collabEditProvider', collabEditProvider);
     this.providerFactory.setProvider('activityProvider', activityProvider);
     this.providerFactory.setProvider('presenceProvider', presenceProvider);