+++ packages/editor-core/src/index.ts
@@ -46,3 +46,4 @@ export { default as EditorContext } from './editor/ui/EditorContext';
 export { default as WithEditorActions } from './editor/ui/WithEditorActions';
 export { default as WithHelpTrigger } from './editor/ui/WithHelpTrigger';
 export { default as getPropsPreset } from './editor/create-editor/get-props-preset';
+export { default as CollapsedEditor } from './editor/ui/CollapsedEditor';