+++ packages/logo/docs/CHANGELOG.md
@@ -1,5 +1,8 @@
 # @atlaskit/logo
 
+## 4.0.2 (2017-10-22)
+
+* bug fix; update styled-components dep and react peerDep ([6a67bf8](https://bitbucket.org/atlassian/atlaskit/commits/6a67bf8))
 ## 4.0.1 (2017-09-14)
 
 * bug fix; jSD and Statuspage logo icons have fixed gradients (issues closed: ak-3479) ([60d8aca](https://bitbucket.org/atlassian/atlaskit/commits/60d8aca))