+++ packages/spinner/docs/CHANGELOG.md
@@ -1,5 +1,8 @@
 # @atlaskit/spinner
 
+## 4.0.1 (2017-10-22)
+
+* bug fix; update styled-components dep and react peerDep ([6a67bf8](https://bitbucket.org/atlassian/atlaskit/commits/6a67bf8))
 ## 4.0.0 (2017-08-29)